+++ src/components/Travel/Item/Modal.tsx
@@ -9,13 +9,10 @@ import {
   ModalFooter,
   ModalBody,
   ModalCloseButton,
-  Button,
   Flex,
 } from '@chakra-ui/react';
 
-import { getLocalStorageItem, setLocalStorageItem } from '@/utils';
-
-import { RESERVATIONS_KEY } from '@/constants/storage';
+import { ReservationButton } from './ReservationButton';
 
 export const TravelItemModal = ({ open, close, itemData }) => {
   const {
@@ -28,18 +25,6 @@ export const TravelItemModal = ({ open, close, itemData }) => {
     maximumPurchases,
     registrationDate,
   } = itemData;
-
-  const handleReservationClick = () => {
-    const reservations = getLocalStorageItem(RESERVATIONS_KEY) || [];
-    const filteredReservations = reservations.filter((reservation) => reservation.idx !== idx);
-    // 로직을 다시 차근차근 세운 후 토스트 알람을 만들기
-    if (filteredReservations.length < reservations) {
-      console.log('장바구니에 이미 담겨있는 상품이에요!');
-    } else {
-      console.log('상품이 장바구니에 잘 담겼어요!');
-    }
-    setLocalStorageItem(RESERVATIONS_KEY, [...filteredReservations, itemData]);
-  };
 
   return (
     <>
@@ -69,9 +54,7 @@ export const TravelItemModal = ({ open, close, itemData }) => {
           </ModalBody>
 
           <ModalFooter>
-            <Button colorScheme='blue' mr={3} onClick={handleReservationClick}>
-              예약하기
-            </Button>
+            <ReservationButton itemData={itemData} />
           </ModalFooter>
         </ModalContent>
       </Modal>
