+++ src/constants/storage.ts
@@ -1,0 +1,1 @@
+export const RESERVATIONS_KEY = 'reservations';
